requirements.txt
django~=4.2.1
djangorestframework~=3.14.0